refactor: Rename package name com.example.demo -> com.example.aiinterviewgen
