+++ src/main/java/com/example/aiinterviewgen/AiInterviewGenApplication.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.example.aiinterviewgen;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
