+++ src/test/java/com/example/aiinterviewgen/AiInterviewGenApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.example.aiinterviewgen;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
